% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlR.R
\docType{package}
\name{controlR-package}
\alias{controlR}
\alias{controlR-package}
\title{controlR}
\description{
A tool for parsing arguments by a so called 'controlfile' to a R-script
}
\details{
The controlfile is used to control the arguments of a R-script. By running the R-script in batch mode,
the user can specify the arguments in the controlfile (without knowledge of R).
In the controlfile the user can specify certain keywords (items) which can be a link to a certain directory, value, date or table..
With the functions \code{get_dir}, \code{get_record} or \code{get_table}, arguments from the controlfile are returned.

Try the examples to get convenient with \pkg{controlR}.
}
\author{
\strong{Maintainer}: Martin Mulder \email{Martin2.Mulder@wur.nl} [copyright holder]

}
