% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{create_label}
\alias{create_label}
\title{Create graph label}
\usage{
create_label(label, ...)
}
\arguments{
\item{label}{character string, label description.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Create graph label
}
\details{
An unit can optionally be added to the label by specifying \code{unit}.
In case \code{linebreak} is set to TRUE the unit is placed underneath the label text (default: FALSE).
}
