% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{create_label_graph}
\alias{create_label_graph}
\title{Create graph label with unit}
\usage{
create_label_graph(label_text, label_unit = NULL)
}
\arguments{
\item{label_text}{character string, label decription.}

\item{label_unit}{character string. unit of \code{label_text}.}
}
\value{
label \code{label_text} with unit \code{label_unit}.
}
\description{
Create graph label with unit
}
\keyword{internal}
