% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{create_title_graph}
\alias{create_title_graph}
\title{Create graph title}
\usage{
create_title_graph(title_text, title_unit = NULL, title_sub = NULL)
}
\arguments{
\item{title_text}{character string, main title.}

\item{title_unit}{character string, unit.}

\item{title_sub}{character string, subtitle.}
}
\value{
label \code{title_text} with unit \code{title_unit} and substitle \code{title_sub}.
}
\description{
Create graph title
}
\keyword{internal}
