% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{dayyear}
\alias{dayyear}
\title{Returns daynumber of year}
\usage{
dayyear(date)
}
\arguments{
\item{date}{a character or numeric vector of suspected dates.}
}
\value{
daynumber of year of \code{date}
}
\description{
Returns daynumber of year
}
\examples{
date <- c("2016-01-01", "2016-12-31", "2019-01-01", "2019-12-31")
dayyear(date)
}
