% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_data}
\alias{download_data}
\title{Download data from URL}
\usage{
download_data(file, url, ...)
}
\arguments{
\item{file}{character string, name of file}

\item{url}{character string, name of url}

\item{...}{further arguments passed to or from other methods}
}
\description{
Download data from URL
}
