% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file_exists}
\alias{file_exists}
\title{Check if file exists}
\usage{
file_exists(file, ...)
}
\arguments{
\item{file}{character string, name of file(s).}

\item{...}{further arguments passed to or from other methods.}
}
\value{
logical
}
\description{
Check if file exists
}
\examples{
# specify file (which exists)
file <- system.file("extdata/control.inp", package = "controlR")
file_exists(file = file)

# in case a file does not exists
file <- system.file("extdata/CoNtRoL.inp", package = "controlR")
file_exists(file = file, case_sensitive = FALSE)
file_exists(file = file, case_sensitive = TRUE)
}
