% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{get_digits}
\alias{get_digits}
\title{Get number of digits of value}
\usage{
get_digits(value)
}
\arguments{
\item{value}{value}
}
\value{
Number of digits of \code{val}
}
\description{
Get number of digits of value
}
\examples{
get_digits(value = 12)
get_digits(value = 12.)
get_digits(value = 12.123)
get_digits(value = c(0.0, 12, 1.23, 1))
}
