% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{get_dir}
\alias{get_dir}
\title{Get directory specified by item}
\usage{
get_dir(item, ...)
}
\arguments{
\item{item}{character string, keyword to specify record.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Character string, name of directory
}
\description{
Get directory from controlfile. The controlfile is specified by either \code{text} or \code{file}.
The directory to extract from the controlfile is specified by keyword \code{item}. Optionally the directory is created (\code{create}).
}
\details{
either \code{text} or \code{file} should be specified.
}
\examples{
# specify controlfile
file <- system.file("extdata/control.inp", package = "controlR")

# get directory specified by 'DIROUT' in controlfile
get_dir(file = file, item = "DIROUT", create = FALSE)
}
