% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{get_line}
\alias{get_line}
\title{Get line in text containing item}
\usage{
get_line(text, item, ...)
}
\arguments{
\item{text}{vector of character strings.}

\item{item}{character string, keyword of which linenumber should be returned.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
integer, line in text with containing \code{item}.
}
\description{
Get line in text containing item
}
\examples{
# specify controlfile
file <- system.file("extdata/control.inp", package = "controlR")
text <- readLines(con = file)

# get line containing 'FILE' in controlfile
get_line(text = text, item = "FILE")
}
\keyword{internal}
