% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_my_theme}
\alias{get_my_theme}
\title{Get theme for plotting}
\usage{
get_my_theme(theme)
}
\arguments{
\item{theme}{character string, theme decription.}
}
\value{
mytheme.
}
\description{
Get theme for plotting
}
