% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{get_record}
\alias{get_record}
\title{Get record specified by item}
\usage{
get_record(item, ...)
}
\arguments{
\item{item}{character string, keyword to specify record.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
character string, date or value specified by \code{item}.
}
\description{
Get record from controlfile. The controlfile is specified by either \code{text} or \code{file}.
The record to extract from the controlfile is specified by keyword \code{item}.
}
\details{
either \code{text} or \code{file} should be specified.
}
\examples{
# specify controlfile
file <- system.file("extdata/control.inp", package = "controlR")

# get record specified by 'FILE' in controlfile
get_record(file = file, item = "FILE")

# get record specified by 'FILE' in controlfile
text <- readLines(con = file)
get_record(text = text, item = "FILE")

# get record specified by 'VALUE' in controlfile
get_record(file = file, item = "VALUE")

# get record specified by 'DATE' in controlfile
get_record(file = file, item = "DATE")
}
