% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{get_table}
\alias{get_table}
\title{Get table specified by item}
\usage{
get_table(item, ...)
}
\arguments{
\item{item}{character string, keyword to specify table.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\code{\link{data.frame}}, table specified by \code{item}.
}
\description{
Get table from controlfile. The controlfile is specified by either \code{text} or \code{file}.
The table to extract from the controlfile is specified by keyword \code{item}.
}
\details{
either \code{text} or \code{file} should be specified.
}
\examples{
# specify controlfile
file <- system.file("extdata/control.inp", package = "controlR")

# get table specified by 'VALUES' in controlfile
get_table(file = file, item = "VALUES", header = "VALUES")
}
