% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{get_variable}
\alias{get_variable}
\title{Get variable(s) from text}
\usage{
get_variable(...)
}
\arguments{
\item{...}{further arguments passed to or from other methods.}
}
\description{
Get variable(s) from text
}
\details{
either \code{text} or \code{file} should be specified.
}
\examples{
# specify controlfile
file <- system.file("extdata/control.inp", package = "controlR")

# collect variable in controlfile
get_variable(file = file)
}
