% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{hydryear}
\alias{hydryear}
\title{Returns hydrologic year}
\usage{
hydryear(date)
}
\arguments{
\item{date}{a character or numeric vector of suspected dates.}
}
\value{
hydrologic year of \code{date}
}
\description{
Returns hydrologic year
}
\examples{
date <- c("2016-01-01", "2016-12-31", "2019-01-01", "2019-12-31")
hydryear(date)
}
