% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{is_date}
\alias{is_date}
\title{Check if string can be converted to date}
\usage{
is_date(text)
}
\arguments{
\item{text}{vector of character strings.}
}
\value{
Logical
}
\description{
Check if string can be converted to date
}
\examples{
is_date(text = "2019-08-13")
is_date(text = "2019-08-13 15:47:00")
is_date(text = "13-08-2019")
is_date(text = "13-Aug-2019")
}
