% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{is_numeric}
\alias{is_numeric}
\title{Check if string can be converted to numerical value}
\usage{
is_numeric(text)
}
\arguments{
\item{text}{vector of character strings.}
}
\value{
Logical
}
\description{
Check if string can be converted to numerical value
}
\examples{
is_numeric(text = "abc")
is_numeric(text = "2.0")
is_numeric(text = "2.0f")
}
