% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{is_windows}
\alias{is_windows}
\title{Check if R is running on a windows machine}
\usage{
is_windows()
}
\value{
Logical
}
\description{
Check if R is running on a windows machine
}
\examples{
is_windows()
}
