% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{sequence_string}
\alias{sequence_string}
\title{Convert numeric sequence to string}
\usage{
sequence_string(sequence)
}
\arguments{
\item{sequence}{numeric sequence.}
}
\value{
Character string
}
\description{
Convert numeric sequence to string
}
\examples{
sequence_string(sequence = c(1,2,3:5,9,11))
sequence_string(sequence = c(1,2,3:5,9,11,"AVG"))
sequence_string(sequence = c(1))
}
