% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{string_sequence}
\alias{string_sequence}
\title{Convert string to numeric sequence}
\usage{
string_sequence(string)
}
\arguments{
\item{string}{character string specifying numeric sequence.}
}
\value{
Numeric vector
}
\description{
Convert string to numeric sequence
}
\examples{
string_sequence(string = "1,2,3-5,10")
string_sequence(string = "1,2,3-5,AVG")
}
