% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{unit_expression}
\alias{unit_expression}
\title{Create unit expression}
\usage{
unit_expression(unit)
}
\arguments{
\item{unit}{character string, unit decription.}
}
\value{
unit expression \code{unit}.
}
\description{
Create unit expression
}
\keyword{internal}
